Enable rule and measure opt-outs (#14)
Rule definitions are mandatory by default, but the rule set may choose to define some rules as optional, or individual measures within a particular rule as optional. In these cases, projects may declare that they these rules or measures do not apply.

Also Improved rigor of domain model. The domain model used to use the same type to represent the definition of a particular type of measure for a rule type, and an actual assertion of that measure in a specific project's IMM. This was initially only mildly confusing, but once we introduced self-aging measures, in which the measure definition looks different from the corresponding IMM entry (the definition has an Age property, the IMM entry has a Date), it became problematic.

Also, as we introduced extra measure types, having one single type to represent everything was not ideal. This PR therefore introduces specific subtypes for Age- and Framework-based rules. It also changes the names to make it clear which types define the rules or measure, and which represent a particular project's assertion that it has a particular set of measures against a particular rule.